requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
cmake==3.26.3
coloredlogs==15.0.1
Django==4.2
djangorestframework==3.14.0
filelock==3.11.0
flatbuffers==23.3.3
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
lit==16.0.1
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnxruntime==1.14.1
packaging==23.1
Pillow==9.5.0
protobuf==4.22.3
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
sympy==1.11.1
torch==2.0.0
torchvision==0.15.1
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
